requirements.txt
fastapi
uvicorn
pydantic
google-adk-agents
google-genai
python-dotenv
fastmcp
httpx
